+++ Bioterio/Models/PlanoAlimentar.cs
@@ -14,9 +14,9 @@ namespace Bioterio.Models
         public string Nome { get; set; }
         public string MarcaAlim { get; set; }
         public int Tipo { get; set; }
-        public double Temperatura { get; set; }
-        public double? Racao { get; set; }
-        public double RacaoDia { get; set; }
+        public float Temperatura { get; set; }
+        public float? Racao { get; set; }
+        public float RacaoDia { get; set; }
 
         public ICollection<RegAlimentar> RegAlimentar { get; set; }
     }
