+++ Bioterio/Models/RegAmostragens.cs
@@ -7,9 +7,9 @@ namespace Bioterio.Models
     {
         public int IdRegAmo { get; set; }
         public DateTime Data { get; set; }
-        public double PesoMedio { get; set; }
+        public float PesoMedio { get; set; }
         public int NroIndividuos { get; set; }
-        public double PesoTotal { get; set; }
+        public float PesoTotal { get; set; }
         public int TanqueIdTanque { get; set; }
 
         public Tanque TanqueIdTanqueNavigation { get; set; }
